chore: move mailbox to be internal
